+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## 1.7.0 - 2021-10-18
+
 ### Added
 
 - Properties `forkingAllowed` and `forkCount` to RepoEntity